pyproject.toml
[project]
name = "hartfordjobs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "flask>=3.1.3",
    "lxml>=6.0.2",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
]
